ListViews now rigged up for the activities that show who the user is monitoring and monitored by.
More attributes were added to the User class, since each User has unique group of people he/she is monitoring and being monitored by, and the corresponding methods to access and modify the said attributes. ArrayAdapters were used to rig the listViews and the User attributes. Button functionality on the the listViews and the activities are coming soon.